requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
Pillow>=10.0.0
django-filter>=24.2
